fix: use tryexecuteasync, jsdoc

+++ src/events/ReadyEvent.ts
@@ -7,6 +7,7 @@ import { HeadcountInstance } from "../instances/HeadcountInstance";
 import { Logger } from "../utilities/Logger";
 import getMongoClient = MongoManager.getMongoClient;
 import { CategoryChannel, GuildTextBasedChannel, TextChannel } from "discord.js";
+import { GlobalFgrUtilities } from "../utilities/fetch-get-request/GlobalFgrUtilities";
 
 const LOGGER: Logger = new Logger(__filename, false);
 
@@ -71,7 +72,7 @@ export async function onReadyEvent(): Promise<void> {
         const fbChannel = Bot.BotInstance.client.channels.cache.get(doc.channels.raids.leaderFeedbackChannelId);
         const storageChannel = Bot.BotInstance.client.channels.cache.get(doc.channels.storageChannelId);
         if (fbChannel?.isText()) {
-            // we know it's in a guild, and we always know parent is category type
+            // we know it's in a guild and text (see above line), and we always know parent is category type
             const parent = (fbChannel as GuildTextBasedChannel).parent as CategoryChannel;
             parent.children.map(channel => {
                 if (!channel.isText()) return;
@@ -79,11 +80,16 @@ export async function onReadyEvent(): Promise<void> {
 
                 const topicUuid = (channel as TextChannel).topic?.split(" ")[0];
                 if (topicUuid && !activeRaidIds.includes(topicUuid)) {
-                    if (storageChannel) {
-                        RaidInstance.compileDeadFeedbackHistory(channel as TextChannel, storageChannel as TextChannel);
-                    } else {
-                        channel.delete();
-                    }
+                    // if no uuid just ignore it, it could be a channel explaining how to give feedback
+                    GlobalFgrUtilities.tryExecuteAsync(async () => {
+                        if (storageChannel && storageChannel.isText()) {
+                            // typecasting is necessary here to remove newschannel possibility. it doesn't affect anything
+                            RaidInstance.compileDeadFeedbackHistory(channel as TextChannel, storageChannel as TextChannel);
+                        } else {
+                            await channel.delete();
+                        }
+                    });
+
                 }
             });
         }
